# openapi/app.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lattis application API",
    "version": "0.1.0-content-draft"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4100"
    }
  ],
  "paths": {
...
    "/api/imports/wordpress/users/cursor": {
      "get": { "summary": "Read WordPress user import cursor", "x-lattis-permission": "user.import:user", "responses": { "200": { "description": "Cursor" }, "401": { "description": "Unauthenticated" }, "403": { "description": "Forbidden" } } }
    },
    "/api/imports/wordpress/users/batches": {
      "post": { "summary": "Import WordPress identities, optionally with encrypted legacy password hashes", "x-lattis-permission": "user.import:user", "responses": { "200": { "description": "Batch imported" }, "400": { "description": "Invalid input" }, "401": { "description": "Unauthenticated" }, "403": { "description": "Forbidden" }, "409": { "description": "Cursor conflict" }, "503": { "description": "Password migration is not configured" } } }
    },
    "/api/me/imported-identities": {
      "get": { "summary": "List source identities for a verified email", "responses": { "200": { "description": "Identities" }, "401": { "description": "Verified account required" } } }
    },
    "/api/me/imported-identities/claim": {
      "post": { "summary": "Claim source identities after email verification", "responses": { "200": { "description": "Claimed count" }, "401": { "description": "Verified account required" }, "409": { "description": "Identity conflict" } } }
    },
...
  },
...
}